int(defence)

+++ MAEnv/scenarios/scenario2_Target.py
@@ -31,7 +31,7 @@ class Scenario(BaseScenario):
             landmark.movable = False
             landmark.value = np.ceil(random.random() * 10)
             landmark.size = landmark.value * 0.01
-            landmark.defence = np.ceil(random.random() * 5)
+            landmark.defence = int(np.ceil(random.random() * 5))
         world.obstacles = [Landmark() for i in range(num_obstacles)]
         for i, landmark in enumerate(world.obstacles):
             landmark.name = 'obstacle %d' % i
